requirements.txt
django==5.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
python-telegram-bot==21.6
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.1.0
psycopg2-binary==2.9.9
APScheduler==3.10.4
requests==2.32.3
